pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "yc-qb-web-connector"
version = "1.0.0"
description = "QuickBooks Desktop Web Connector → Supabase sync for Nature's Storehouse & ADK Fragrance Farm"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "spyne>=2.14",
    "lxml>=5.1",
    "supabase>=2.4",
    "pydantic-settings>=2.2",
    "structlog>=24.1",
    "tenacity>=8.2",
    "pyyaml>=6.0",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.4",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
log_cli = true
log_cli_level = "WARNING"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "W"]
ignore = ["E501"]
